Fix a bunch of CMake compilation warnings
Bug: 436300898
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/287276
Commit-Queue: Shao, Jiawei <jiawei.shao@intel.com>

+++ src/dawn/common/Math.cpp
@@ -104,7 +104,7 @@ uint16_t Float32ToFloat16(float fp32) {
 float Float16ToFloat32(uint16_t fp16) {
     uint32_t tmp = (fp16 & 0x7fff) << 13 | (fp16 & 0x8000) << 16;
     float tmp2 = *reinterpret_cast<float*>(&tmp);
-    return pow(2, 127 - 15) * tmp2;
+    return powf(2.0f, 127.0f - 15.0f) * tmp2;
 }
 
 bool IsFloat16NaN(uint16_t fp16) {
