+++ src/dawn/native/RenderEncoderBase.cpp
@@ -742,11 +742,11 @@ void RenderEncoderBase::APISetImmediates(uint32_t offset, const void* data, size
 
             SetImmediatesCmd* cmd = allocator->Allocate<SetImmediatesCmd>(Command::SetImmediates);
             cmd->offset = offset;
-            cmd->size = size;
+            cmd->size = uint32_t(size);
             uint8_t* immediateDatas = allocator->AllocateData<uint8_t>(cmd->size);
             memcpy(immediateDatas, data, size);
 
-            mCommandBufferState.SetImmediateData(offset, size);
+            mCommandBufferState.SetImmediateData(offset, uint32_t(size));
 
             return {};
         },
